+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/registry/PolarisAutoServiceRegistration.java
@@ -17,11 +17,11 @@
 
 package com.tencent.cloud.polaris.registry;
 
+import com.tencent.cloud.polaris.registry.graceful.AbstractGracefulServiceRegistration;
+import com.tencent.cloud.polaris.registry.graceful.GracefulServiceRegistrationProperties;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import org.springframework.cloud.client.serviceregistry.AbstractAutoServiceRegistration;
-import org.springframework.cloud.client.serviceregistry.AutoServiceRegistrationProperties;
 import org.springframework.cloud.client.serviceregistry.Registration;
 import org.springframework.cloud.client.serviceregistry.ServiceRegistry;
 import org.springframework.util.StringUtils;
@@ -29,16 +29,16 @@ import org.springframework.util.StringUtils;
 /**
  * Auto service registration of Polaris.
  *
- * @author Haotian Zhang, Andrew Shan, Jie Cheng
+ * @author Haotian Zhang, Andrew Shan, Jie Cheng, cheese8
  */
-public class PolarisAutoServiceRegistration extends AbstractAutoServiceRegistration<Registration> {
+public class PolarisAutoServiceRegistration extends AbstractGracefulServiceRegistration<Registration> {
 
 	private static final Logger LOG = LoggerFactory.getLogger(PolarisAutoServiceRegistration.class);
 
 	private final PolarisRegistration registration;
 
 	public PolarisAutoServiceRegistration(ServiceRegistry<Registration> serviceRegistry,
-			AutoServiceRegistrationProperties autoServiceRegistrationProperties,
+			GracefulServiceRegistrationProperties autoServiceRegistrationProperties,
 			PolarisRegistration registration) {
 		super(serviceRegistry, autoServiceRegistrationProperties);
 		this.registration = registration;