+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/registry/graceful/GracefulServiceRegistration.java
@@ -1,0 +1,5 @@
+package com.tencent.cloud.polaris.registry.graceful;
+
+public interface GracefulServiceRegistration {
+
+}