+++ spring-cloud-starter-tencent-polaris-discovery/src/test/java/com/tencent/cloud/polaris/registry/PolarisServiceRegistryTest.java
@@ -43,7 +43,7 @@ import static org.mockito.Mockito.when;
 /**
  * Test for {@link PolarisServiceRegistry}.
  *
- * @author Haotian Zhang
+ * @author Haotian Zhang, cheese8
  */
 public class PolarisServiceRegistryTest {
 
@@ -57,6 +57,7 @@ public class PolarisServiceRegistryTest {
 					PolarisDiscoveryAutoConfiguration.class))
 			.withPropertyValues("spring.application.name=" + SERVICE_PROVIDER)
 			.withPropertyValues("server.port=" + PORT)
+			.withPropertyValues("spring.cloud.service-registry.graceful-registration.enabled=true")
 			.withPropertyValues("spring.cloud.polaris.address=grpc://127.0.0.1:10081")
 			.withPropertyValues("spring.cloud.polaris.discovery.namespace=" + NAMESPACE_TEST)
 			.withPropertyValues("spring.cloud.polaris.discovery.token=xxxxxx");